requirements.txt
absl-py==0.15.0
appdirs==1.4.3
apturl==0.5.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==5.3.1
certifi==2019.11.28
chardet==3.0.4
click==8.1.6
colorama==0.4.3
comm==0.1.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
distlib==0.3.7
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
executing==1.2.0
fasteners==0.14.1
filelock==3.12.2
Flask==2.0.1
flatbuffers==1.12
fugashi==1.1.0
future==0.18.2
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.12.2
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jupyter-client==8.3.0
jupyter-core==5.3.1
jupyterlab-widgets==3.0.9
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
monotonic==1.5
more-itertools==4.2.0
nest-asyncio==1.5.6
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==23.1
paramiko==2.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==3.9.1
progressbar==2.5
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycups==1.9.73
Pygments==2.15.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.0
regex==2023.6.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
sacremoses==0.0.53
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
ssh-import-id==5.10
stack-data==0.6.2
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tokenizers==0.10.3
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.5.1
typing-extensions==3.7.4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidic-lite==1.0.8
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.24.2
wadllib==1.3.3
wcwidth==0.2.6
Werkzeug==2.3.6
widgetsnbextension==4.0.9
wrapt==1.12.1
xkit==0.0.0
zipp==3.16.2
